convert PrettyPrinter to typescript
